third_party/WebKit/LayoutTests/fast/css/zoom-body-scroll-expected.txt

Test for rdar://problem/6643103 Unscaled values from body.scrollHeight.

scrollHeight: 1000
scrollWidth: 1000

Scrolling right to 50
scrollLeft: 50

Zooming in
scrollHeight: 1000
scrollWidth: 1000
scrollTop: 0
scrollLeft: 49

Scrolling down to 100
scrollTop: 99

Zooming back out
scrollTop: 100
scrollLeft: 50
